Refactor: move specifc `getPopoverArrowClasses` helper
Inside the core `Popover` component as to be self-contained
