+++ src/utils/css/index.js
@@ -1,4 +1,5 @@
 /* @flow */
+/* eslint-disable import/prefer-default-export */
 
 import { capitalize } from '../strings';
 
@@ -42,42 +43,4 @@ export const getMainClasses = (
     .map(key => (state[key] ? styleObject[`state${capitalize(key)}`] : ''))
     .filter(i => !!i);
   return [...styleArray, ...modifierClasses, ...stateClasses].join(' ');
-};
-
-/**
- * This is a sripped down version of the above method, that is specifically
- * designed to generate the classes required to properly render the Popover
- * Component's various arrows.
- *
- * @NOTE It won't work anywhere else, and in all fairness it should be placed
- * inside the actual component, but to preserve code cleanliness, I've placed
- * it here.
- *
- * @method getPopoverArrowClasses
- *
- * @param {object} appearance Appearance object
- * @param {placement} string Positioning of the arrow (this is reversed from in the Popover)
- * @param {styleObject} CSS modules styles object
- *
- * @return {[type]} [description]
- */
-export const getPopoverArrowClasses = (
-  { theme }: Object = {},
-  placement: string,
-  styleObject: { [string]: string } = {},
-) => {
-  /*
-   * @NOTE Arrows have the position encoded in the class's name
-   */
-  const placementClass = `${placement}Arrow`;
-  const styleArray = [styleObject[placementClass]];
-  if (theme) {
-    /*
-     * @NOTE Arrows have the theme name encoded in the class's name (which also
-     * has the position encoded, as above)
-     */
-    const themeClass = `theme${capitalize(theme)}${capitalize(placementClass)}`;
-    styleArray.push(styleObject[themeClass]);
-  }
-  return styleArray.join(' ');
 };
